requirements.txt
asgiref==3.5.2
# backports.zoneinfo==0.2.1
click==8.0.3
Django==3.2
django-autocomplete-light==3.8.2
django-filebrowser-no-grappelli==3.8.0
django-nested-admin==3.4.0
django-tabbed-admin==1.0.4
django-tables2==2.4.1
filebrowser==1.1.3
mysqlclient==2.1.1
Pillow==8.4.0
python-monkey-business==1.0.0
pytz==2021.3
six==1.16.0
django-filter==22.1
# django-mptt==0.8.7
# django-polymorphic==1.0.2
easy-thumbnails==2.8.5
olefile==0.46
Unidecode==0.4.21
# django-debug-toolbar==1.7
appdirs==1.4.3
django_environ==0.9

